requirements.txt
swig
pyscard
pandas
openpyxl
kivy
pygobject
pycairo
Pillow
requests